+++ src/main/java/se/bjurr/sbcc/SbccUserAdminService.java
@@ -4,4 +4,6 @@ public interface SbccUserAdminService {
   boolean emailExists(String emailAddress);
 
   boolean displayNameExists(String name);
+
+  boolean slugExists(String name);
 }